refactor(gui): rename system_tools module to custom_tools
Aligns the Tauri backend module name with the frontend's
customSystemTools.ts naming, since this module backs the
selectable/optional "custom system tools" (currently the HTTP
GET test tool) rather than always-on built-ins.
