+++ crates/agent-gui/src-tauri/src/commands/app/custom_tools.rs
