+++ crates/agent-gui/src-tauri/src/commands/app/mod.rs
@@ -1,4 +1,4 @@
 pub mod app;
+pub mod custom_tools;
 pub mod system;
-pub mod system_tools;
 pub mod update;
