+++ crates/agent-gui/src-tauri/src/commands/mod.rs
@@ -14,8 +14,8 @@ pub mod runtime_commands;
 pub mod workspace_commands;
 
 pub use app_commands::app;
+pub use app_commands::custom_tools;
 pub use app_commands::system;
-pub use app_commands::system_tools;
 pub use app_commands::update;
 
 pub use automation_commands::cron;
