+++ crates/agent-gui/src-tauri/src/lib.rs
@@ -217,7 +217,7 @@ macro_rules! app_invoke_handler {
             commands::system::system_append_debug_jsonl,
             commands::system::system_begin_power_activity,
             commands::system::system_end_power_activity,
-            commands::system_tools::system_http_get_test,
+            commands::custom_tools::system_http_get_test,
             commands::gateway::gateway_connect,
             commands::gateway::gateway_disconnect,
             commands::gateway::gateway_status,
